Branching Assignment
This code asks the user to input the values of the package weight, height, length and width. If the use weight is larger than 50 pounds the program issues a statement to the user and terminates. If it is under, the program will gather the other dimensions of the package and compute the cost of shipping.