Console App Array Submission
Create a one-dimensional Array of strings. Ask the user to select an index of the Array and then display the string at that index on the screen.

Create a one-dimensional Array of integers. Ask the user to select an index of the Array and then display the integer at that index on the screen.

Add in a message that displays when the user selects an index that doesn’t exist.

Create a list of strings. Ask the user to select an index of the list and then display the content at that index on the screen.

Add comments to each line or block of your code to explain what it does exactly, so that another developer could read and understand your code.